list make

+++ models/order.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Order struct {
-	ID           uint `json:"id" gorm:"primaryKey"`
+	ID           uint64 `json:"id" gorm:"primaryKey"`
 	CreatedAt    time.Time
 	ProductRefer int     `json:"product_id"`
 	Product      Product `gorm:"foreignKey:ProductRefer"`
